+++ mockwebserver/src/main/java/okhttp3/mockwebserver/internal/tls/HeldCertificate.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package okhttp3.internal.tls;
+package okhttp3.mockwebserver.internal.tls;
 
 import java.math.BigInteger;
 import java.security.GeneralSecurityException;
