+++ okhttp-tests/src/test/java/okhttp3/internal/tls/ClientAuthTest.java
@@ -20,7 +20,6 @@ import java.net.SocketException;
 import java.security.GeneralSecurityException;
 import java.security.NoSuchAlgorithmException;
 import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLException;
 import javax.net.ssl.SSLHandshakeException;
 import javax.net.ssl.SSLPeerUnverifiedException;
 import javax.net.ssl.SSLSocket;
@@ -33,6 +32,8 @@ import okhttp3.Request;
 import okhttp3.Response;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
+import okhttp3.mockwebserver.internal.tls.HeldCertificate;
+import okhttp3.mockwebserver.internal.tls.SslClient;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
